+++ tests/examples/render_jinja_python/variables.py
@@ -1,0 +1,4 @@
+from stepup.core.api import getenv
+
+ENV_VAR_TEST_STEPUP_RENDER = getenv("ENV_VAR_TEST_STEPUP_RENDER", "word")
+title = "Made with RepRep"