+++ tests/examples/render_jinja_relpath/variables.py
@@ -1,0 +1,4 @@
+from stepup.core.api import getenv
+
+ROOT = getenv("ROOT", back=True)
+PUBLIC = getenv("PUBLIC", back=True)